Create README.md
AfterCarrey is a PHP and MySQL web application built with the WAMP stacks. It is an appointment booking web app that allows users to register as patients to set appointments with doctors who are accessible to provide users with relevant aftercare services based on their unique health issues.



Users may view each doctor's available time slots, so they know when the doctors are available for reservations. Furthermore, unregistered users or visitors can send a message to the Localhost with any questions or issues they may have.



The application also has an admin system that allows staff (doctors or administrators) to create, read, update, and delete the following categories: Appointment, Messages, Doctors, Medical Specialties, Users, and Site Settings.